Fix Hough transform and move rasterization
Hough line transform had incorrect
rounding which lead to misleading message
