+++ include/boost/gil.hpp
@@ -39,9 +39,9 @@
 #include <boost/gil/position_iterator.hpp>
 #include <boost/gil/premultiply.hpp>
 #include <boost/gil/promote_integral.hpp>
-#include <boost/gil/rasterization/circle.hpp>
-#include <boost/gil/rasterization/ellipse.hpp>
-#include <boost/gil/rasterization/line.hpp>
+#include <boost/gil/extension/rasterization/circle.hpp>
+#include <boost/gil/extension/rasterization/ellipse.hpp>
+#include <boost/gil/extension/rasterization/line.hpp>
 #include <boost/gil/rgb.hpp>
 #include <boost/gil/rgba.hpp>
 #include <boost/gil/step_iterator.hpp>
