+++ include/boost/gil/extension/rasterization/ellipse.hpp
