+++ test/core/CMakeLists.txt
@@ -40,4 +40,3 @@ add_subdirectory(image_view)
 add_subdirectory(algorithm)
 add_subdirectory(image_processing)
 add_subdirectory(histogram)
-add_subdirectory(rasterization)
